lmkd: add one task in proc_get_script per polling interval
In search of native apps to kill by ULMK, proc_get_script() traverse all
the process in the system by opening /proc directory. This can take
indefinte amount of time in low on memory as to give the info for each
entry kernel may go for memory reclaim. As anyway we are killing only
one process in a polling interval, just limit adding one eligible process
to kill while saving the current position in readdir.

CRs-Fixed: 2521954
Change-Id: I94c8a1c61735d1412c7efc83e9d627730ac2c06b